+++ bot/src/main/kotlin/dev/inmo/plagubot/HelloPlugin.kt
@@ -1,10 +1,18 @@
 package dev.inmo.plagubot
 
 import dev.inmo.kslog.common.*
+import dev.inmo.micro_utils.fsm.common.State
+import dev.inmo.plagubot.HelloPlugin.setupBotPlugin
 import dev.inmo.tgbotapi.extensions.api.bot.getMe
 import dev.inmo.tgbotapi.extensions.api.send.reply
-import dev.inmo.tgbotapi.extensions.behaviour_builder.BehaviourContext
+import dev.inmo.tgbotapi.extensions.api.send.sendMessage
+import dev.inmo.tgbotapi.extensions.behaviour_builder.*
+import dev.inmo.tgbotapi.extensions.behaviour_builder.expectations.waitText
+import dev.inmo.tgbotapi.extensions.behaviour_builder.expectations.waitTextMessage
 import dev.inmo.tgbotapi.extensions.behaviour_builder.triggers_handling.onCommand
+import dev.inmo.tgbotapi.extensions.behaviour_builder.triggers_handling.onUnhandledCommand
+import dev.inmo.tgbotapi.types.ChatId
+import kotlinx.coroutines.flow.first
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
@@ -24,14 +32,34 @@ object HelloPlugin : Plugin {
     override fun Module.setupDI(database: Database, params: JsonObject) {
         single {
             get<Json>().decodeFromJsonElement(HelloPluginConfig.serializer(), params["helloPlugin"] ?: return@single null)
+
+
         }
     }
 
-    override suspend fun BehaviourContext.setupBotPlugin(koin: Koin) {
-        logger.d { koin.get<HelloPluginConfig>().print }
+    private sealed interface InternalFSMState : State {
+        override val context: ChatId
+        data class DidntSaidHello(override val context: ChatId) : InternalFSMState
+        data class SaidHelloOnce(override val context: ChatId) : InternalFSMState
+    }
+
+    override suspend fun BehaviourContextWithFSM<State>.setupBotPlugin(koin: Koin) {
+        val toPrint = koin.getOrNull<HelloPluginConfig>() ?.print ?: "Hello :)"
+        logger.d { toPrint }
         logger.dS { getMe().toString() }
         onCommand("hello_world") {
-            reply(it, "Hello :)")
+            startChain(InternalFSMState.DidntSaidHello(it.chat.id))
+        }
+
+        strictlyOn { state: InternalFSMState.DidntSaidHello ->
+            sendMessage(state.context, toPrint)
+            InternalFSMState.SaidHelloOnce(state.context)
+        }
+
+        strictlyOn { state: InternalFSMState.SaidHelloOnce ->
+            val message = waitTextMessage().first()
+            reply(message, "Sorry, I can answer only this: $toPrint")
+            InternalFSMState.SaidHelloOnce(state.context)
         }
     }
 }